refactor(game): collapse renderer-backend layers (PR 2)
Merge V2 + compat shim into a single canonical RendererBackend interface
and delete the parallel files. The v2 interface declared every method
optional and the compat module existed only to `??` past those optionals,
which stalled the migration and left every call site guarded by
intersection types.

- Unify types in renderer-backend.ts; make applyEnvironment,
  applyPostProcessPlan, applyQuality, dispose, initialize, renderFrame,
  and resize required.
- Delete renderer-backend-v2.ts, renderer-backend-compat.ts, and their
  tests; repoint the 30 consumers at renderer-backend.
- Simplify the loader and runtime now that the compat shim is gone.
- Give the WebGPU backend a no-op applyEnvironment so it satisfies the
  required contract without pretending to run PMREM/HDR.
- Hoist renderer-backend-runtime test fakes with vi.hoisted so the
  importOriginal-backed mock stops racing @/ui/config.

Verification:
- rg "RendererBackendV2|backend-compat|RendererBackendCompat" is empty.
- src/three vitest suite: 374 files / 1622 tests passing.
- pnpm tsc --noEmit -p tsconfig.json clean.
