+++ client/apps/game/src/three/managers/point-label-texture-policy.ts
@@ -1,4 +1,4 @@
-import type { RendererActiveMode } from "../renderer-backend-v2";
+import type { RendererActiveMode } from "../renderer-backend";
 
 export function resolvePointLabelTextureFlipY(activeMode: RendererActiveMode | null): boolean {
   return activeMode === "legacy-webgl";
