+++ client/apps/game/src/three/renderer-destroy-runtime.ts
@@ -1,6 +1,4 @@
-import { disposeRendererBackend } from "./renderer-backend-compat";
-import type { RendererSurfaceLike } from "./renderer-backend";
-import type { RendererBackendV2 } from "./renderer-backend-v2";
+import type { RendererBackend, RendererSurfaceLike } from "./renderer-backend";
 import type { RendererInteractionRuntime } from "./renderer-interaction-runtime";
 import type { RendererLabelRuntime } from "./renderer-label-runtime";
 import type { RendererMonitoringRuntime } from "./renderer-monitoring-runtime";
@@ -14,7 +12,7 @@ type Destroyable = {
 };
 
 interface DestroyRendererRuntimeInput {
-  backend?: RendererBackendV2 & { renderer: RendererSurfaceLike; dispose?: () => void };
+  backend?: RendererBackend;
   cleanupIntervals: NodeJS.Timeout[];
   controls?: { dispose(): void };
   effectsBridgeRuntime?: Pick<RendererEffectsBridgeRuntime, "dispose">;
@@ -59,14 +57,8 @@ function detachRendererSurface(renderer?: RendererSurfaceLike): void {
   }
 }
 
-function disposeRendererResources(
-  backend?: RendererBackendV2 & { renderer: RendererSurfaceLike; dispose?: () => void },
-): void {
-  if (!backend) {
-    return;
-  }
-
-  disposeRendererBackend(backend);
+function disposeRendererResources(backend?: RendererBackend): void {
+  backend?.dispose();
 }
 
 function destroyRendererScenes(scenes: DestroyRendererRuntimeInput["scenes"]): void {
