+++ client/apps/game/src/three/three-webgpu-type-guard.ts
@@ -1,13 +1,13 @@
 import type { WebGPURenderer } from "three/webgpu";
 
-import type { RendererBackendV2 } from "./renderer-backend-v2";
+import type { RendererBackend } from "./renderer-backend";
 
 type AssertTrue<T extends true> = T;
 type AssertFalse<T extends false> = T;
 type IsAny<T> = 0 extends 1 & T ? true : false;
 
 type WebGpuRendererType = WebGPURenderer;
-type BackendCapabilitiesType = RendererBackendV2["capabilities"];
+type BackendCapabilitiesType = RendererBackend["capabilities"];
 
 type _webGpuRendererIsTyped = AssertFalse<IsAny<WebGpuRendererType>>;
 type _backendCapabilitiesAreTyped = AssertFalse<IsAny<BackendCapabilitiesType>>;
