+++ client/apps/game/src/three/webgpu-postprocess-policy.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 
-import { createRendererBackendCapabilities } from "./renderer-backend-v2";
+import { createRendererBackendCapabilities } from "./renderer-backend";
 import { resolveWebgpuPostprocessPolicy } from "./webgpu-postprocess-policy";
 import { requestRendererScenePrewarm } from "./webgpu-postprocess-policy";
