+++ client/apps/game/src/three/webgpu-postprocess-runtime.test.ts
@@ -2,7 +2,7 @@ import { ACESFilmicToneMapping, NeutralToneMapping } from "three";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { createWebGPUPostProcessRuntime } from "./webgpu-postprocess-runtime";
-import type { RendererPostProcessPlan } from "./renderer-backend-v2";
+import type { RendererPostProcessPlan } from "./renderer-backend";
 
 function createPlan(overrides: Partial<RendererPostProcessPlan> = {}): RendererPostProcessPlan {
   return {
